requirements.txt
pip
discord.py
psycopg2
